+++ lending/RELEASES.md
@@ -173,3 +173,13 @@ Based on:
 - [csharp v5.1.0] lending
 ### Releases
 - [NuGet v5.1.0] https://www.nuget.org/packages/Codat.Lending/5.1.0 - lending
+
+## 2023-12-15 12:55:54
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.126.3 (2.214.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v5.2.0] lending
+### Releases
+- [NuGet v5.2.0] https://www.nuget.org/packages/Codat.Lending/5.2.0 - lending